feat: 사용자의 주차장 반환 API 및 서비스 메소드의 위치를 User 컨트롤러와 User 서비스에 위치하게 변경
